fix(IX): fixed issues and refactored code
