+++ scripts/MC.py
@@ -1,7 +1,7 @@
 from time import sleep
 from helpers.clientFinder.dataLookup import dataLookup
-from helpers.operations.addHandler import addOnuService
-from helpers.operations.newAddHandler import addOnuServiceNew
+# from helpers.operations.addHandler import addOnuService
+from helpers.operations.addHandler import addOnuServiceNew
 from helpers.operations.spid import availableSpid, spidCalc, verifySPID
 from helpers.utils.display import display
 from helpers.utils.printer import colorFormatter, inp, log
@@ -24,98 +24,82 @@ $ """
     )
     lookupType = inp(
         "Buscar cliente por serial o por Datos (F/S/P/ID) [S | D] : ")
-    data = dataLookup(comm, command, olt, lookupType)
-    proceed = display(data, "A")
+    client = dataLookup(comm, command, olt, lookupType)
+    proceed = display(client, "A")
     if not proceed:
         log(colorFormatter("Cancelando...", "warning"))
         quit()
         return
     if action == "CT":
         NEW_NAME = inp("Ingrese el Nuevo Titular del cliente : ")
-        command(f"interface gpon {data['frame']}/{data['slot']}")
-        command(f'ont modify {data["port"]} {data["id"]} desc "{NEW_NAME}" ')
+        command(f"interface gpon {client['frame']}/{client['slot']}")
+        command(f'ont modify {client["port"]} {client["id"]} desc "{NEW_NAME}" ')
         command("quit")
         log(
             colorFormatter(
-                f"Al cliente {data['name']} {data['frame']}/{data['slot']}/{data['port']}/{data['id']} @ OLT {data['olt']} se le ha cambiado el nombre a '{NEW_NAME}'",
+                f"Al cliente {client['name']} {client['frame']}/{client['slot']}/{client['port']}/{client['id']} @ OLT {client['olt']} se le ha cambiado el nombre a '{NEW_NAME}'",
                 "success",
             )
         )
-        modify(data["sn"], NEW_NAME, "NAME")
+        modify(client["sn"], NEW_NAME, "NAME")
         quit()
         return
     if action == "CO":
         NEW_SN = inp("Ingrese el Nuevo serial de ONT del cliente : ")
-        command(f"interface gpon {data['frame']}/{data['slot']}")
-        command(f'ont modify {data["port"]} {data["id"]} sn "{NEW_SN}" ')
+        command(f"interface gpon {client['frame']}/{client['slot']}")
+        command(f'ont modify {client["port"]} {client["id"]} sn "{NEW_SN}" ')
         command("quit")
         log(
             colorFormatter(
-                f"Al cliente {data['name']} {data['frame']}/{data['slot']}/{data['port']}/{data['id']} @ OLT {data['olt']} se le ha cambiado el serial a '{NEW_SN}'",
+                f"Al cliente {client['name']} {client['frame']}/{client['slot']}/{client['port']}/{client['id']} @ OLT {client['olt']} se le ha cambiado el serial a '{NEW_SN}'",
                 "success",
             )
         )
-        modify(data["sn"], NEW_SN, "SN")
+        modify(client["sn"], NEW_SN, "SN")
         quit()
         return
     if action == "CP":
-        if olt != "1":
-            for wan in data["wan"]:
-                command(f"undo service-port {wan['spid']}")
-            addOnuService(comm,command,data)
-        else:
-            data['planName'] = inp("Ingrese el Nuevo plana instalar : ")
-            data['wan'] = plans[data['planName']]
-            for wanData in data["wan"]:
-                command(f"undo service-port {wanData['spid']}")
-            addOnuServiceNew(comm,command,data)
-            verifySPID(comm, command, data)
+        client['planName'] = inp("Ingrese el Nuevo plana instalar : ")
+        client['wan'] = plans[client['planName']]
+        for wanData in client["wan"]:
+            command(f"undo service-port {wanData['spid']}")
+        addOnuServiceNew(comm,command,client)
+        verifySPID(comm, command, client)
         log(
             colorFormatter(
-                f"Al cliente {data['name']} {data['frame']}/{data['slot']}/{data['port']}/{data['id']} @ OLT {data['olt']} se le ha Cambiado el plan y vlan a {data['planName']} @ {data['wan']['vlan']}",
+                f"Al cliente {client['name']} {client['frame']}/{client['slot']}/{client['port']}/{client['id']} @ OLT {client['olt']} se le ha Cambiado el plan y vlan a {client['planName']} @ {client['wan']['vlan']}",
                 "info",
             )
         )
-        modify(data["sn"], data['planName'], "PLAN")
-        modify(data["sn"], data['wan']['vlan'], "PROVIDER")
+        modify(client["sn"], client['planName'], "PLAN")
+        modify(client["sn"], client['wan']['vlan'], "PROVIDER")
         quit()
         return
     if action == "ES":
         log("| {:^3} | {:^4} | {:^6} |".format("IDX", "VLAN", "SPID"))
-        for idx, wan in enumerate(data["wan"]):
+        for idx, wan in enumerate(client["wan"]):
             log("| {:^3} | {:^4} | {:^6} |".format(
                 idx, wan["vlan"], wan["spid"]))
         DEL_SPID = int(inp("Ingrese el SPID a eliminar : "))
         command(f"undo service-port {DEL_SPID}")
         log(
             colorFormatter(
-                f"Al cliente {data['name']} {data['frame']}/{data['slot']}/{data['port']}/{data['id']} @ OLT {data['olt']} se le ha eliminado el SPID {DEL_SPID}",
+                f"Al cliente {client['name']} {client['frame']}/{client['slot']}/{client['port']}/{client['id']} @ OLT {client['olt']} se le ha eliminado el SPID {DEL_SPID}",
                 "info",
             )
         )
         quit()
         return
     if action == "AS":
-        if olt != "1":
-            data["wan"][0]["spid"] = availableSpid(comm, command)
-            addOnuService(comm, command, data)
-            verifySPID(comm, command, data)
-            log(
-                colorFormatter(
-                    f"Al cliente {data['name']} {data['frame']}/{data['slot']}/{data['port']}/{data['id']} @ OLT {data['olt']} se le ha Agregado el plan y vlan a {data['plan']} @ {data['vlan']}",
-                    "info",
-                )
-            )
-            quit()
-            return
         NEW_PLAN = inp("Ingrese el Nuevo Plan del cliente : ")
-        data["wan"][0]["vlan"] = plans[NEW_PLAN]["vlan"]
-        data["wan"][0]["plan"] = plans[NEW_PLAN]["plan"]
-        data["gemPort"] = plans[NEW_PLAN]["gemPort"]
-        addOnuServiceNew(comm, command, data)
+        client["wan"][0]["vlan"] = plans[NEW_PLAN]["vlan"]
+        client["wan"][0]["plan"] = plans[NEW_PLAN]["plan"]
+        client["gemPort"] = plans[NEW_PLAN]["gemPort"]
+        addOnuServiceNew(comm, command, client)
+        verifySPID(comm, command, client)
         log(
             colorFormatter(
-                f"Al cliente {data['name']} {data['frame']}/{data['slot']}/{data['port']}/{data['id']} @ OLT {data['olt']} se le ha Agregado el plan y vlan a {NEW_PLAN} @ {data['wan'][0]['vlan']}",
+                f"Al cliente {client['name']} {client['frame']}/{client['slot']}/{client['port']}/{client['id']} @ OLT {client['olt']} se le ha Agregado el plan y vlan a {NEW_PLAN} @ {client['wan'][0]['vlan']}",
                 "info",
             )
         )
@@ -123,8 +107,8 @@ $ """
         return
         
     if action == "AV":
-        data["wan"][0]["spid"] = spidCalc(data)["V"]
+        client["wan"][0]["spid"] = spidCalc(client)["V"]
         log(colorFormatter(
-            f"SPID PARA AGG {data['spid']}, FUNCION AUN NO DISPONIBLE"), "info")
+            f"SPID PARA AGG {client['spid']}, FUNCION AUN NO DISPONIBLE"), "info")
         quit()
         return
